+++ package.json
@@ -16,7 +16,7 @@
     "react-scripts": "5.0.1",
     "styled-components": "^6.1.1",
     "styled-reset": "^4.5.1",
-    "typescript": "^4.9.5",
+    "typescript": "5.1.6",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
