--- src/common/ui/button/Button.stories.tsx
@@ -1,18 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import Button from "./Buttons";
-
-const meta: Meta<typeof Button> = {
-  title: "UI/button/Button",
-  component: Button,
-};
-
-export default meta;
-type Story = StoryObj<typeof Button>;
-
-export const ButtonLayout: Story = {
-  args: {
-    width: "20rem",
-    height: "7rem",
-    bgColor: "skyblue",
-  },
-};
